Bubbles: - Renamed BubblesNotificationView to just Bubbles (like all other views)
Bezel:
- Use constant for preference domain

Smoke:
- Minor changes

MusicVideo:
- Forgot to import GrowlStringAdditions

+++ Display Plugins/Bezel/GrowlBezelWindowController.h
@@ -16,7 +16,7 @@
 	SEL				_action;
 	id				_target;
 	id				_representedObject;
-	short			_displayTime;
+	double			_displayTime;
 	int				_priority;
 }
